积累更新

+++ NativeCode/Base/perlin/perlin.cpp
@@ -27,10 +27,10 @@ NS_BEGIN
 	r0 = t - (int)t;\
 	r1 = r0 - 1.0f;
 
-float Perlin::noise1(float arg)
+double Perlin::noise1(double arg)
 {
 	int bx0, bx1;
-	float rx0, rx1, sx, t, u, v, vec[1];
+	double rx0, rx1, sx, t, u, v, vec[1];
 
 	vec[0] = arg;
 
@@ -51,10 +51,10 @@ float Perlin::noise1(float arg)
 	return lerp(sx, u, v);
 }
 
-float Perlin::noise2(float vec[2])
+double Perlin::noise2(double vec[2])
 {
 	int bx0, bx1, by0, by1, b00, b10, b01, b11;
-	float rx0, rx1, ry0, ry1, *q, sx, sy, a, b, t, u, v;
+	double rx0, rx1, ry0, ry1, *q, sx, sy, a, b, t, u, v;
 	int i, j;
 
 	if (mStart)
@@ -95,10 +95,10 @@ float Perlin::noise2(float vec[2])
 	return lerp(sy, a, b);
 }
 
-float Perlin::noise3(float vec[3])
+double Perlin::noise3(double vec[3])
 {
 	int bx0, bx1, by0, by1, bz0, bz1, b00, b10, b01, b11;
-	float rx0, rx1, ry0, ry1, rz0, rz1, *q, sy, sz, a, b, c, d, t, u, v;
+	double rx0, rx1, ry0, ry1, rz0, rz1, *q, sy, sz, a, b, c, d, t, u, v;
 	int i, j;
 
 	if (mStart)
@@ -149,21 +149,21 @@ float Perlin::noise3(float vec[3])
 	return lerp(sz, c, d);
 }
 
-void Perlin::normalize2(float v[2])
+void Perlin::normalize2(double v[2])
 {
-	float s;
+	double s;
 
-	s = (float)sqrtf(v[0] * v[0] + v[1] * v[1]);
+	s = sqrt(v[0] * v[0] + v[1] * v[1]);
 	s = 1.0f/s;
 	v[0] = v[0] * s;
 	v[1] = v[1] * s;
 }
 
-void Perlin::normalize3(float v[3])
+void Perlin::normalize3(double v[3])
 {
-	float s;
+	double s;
 
-	s = (float)sqrtf(v[0] * v[0] + v[1] * v[1] + v[2] * v[2]);
+	s = sqrt(v[0] * v[0] + v[1] * v[1] + v[2] * v[2]);
 	s = 1.0f/s;
 
 	v[0] = v[0] * s;
@@ -178,12 +178,12 @@ void Perlin::init(void)
 	for (i = 0 ; i < B ; i++)
 	{
 		p[i] = i;
-		g1[i] = (float)((rand() % (B + B)) - B) / B;
+		g1[i] = (double)((rand() % (B + B)) - B) / B;
 		for (j = 0 ; j < 2 ; j++)
-			g2[i][j] = (float)((rand() % (B + B)) - B) / B;
+			g2[i][j] = (double)((rand() % (B + B)) - B) / B;
 		normalize2(g2[i]);
 		for (j = 0 ; j < 3 ; j++)
-			g3[i][j] = (float)((rand() % (B + B)) - B) / B;
+			g3[i][j] = (double)((rand() % (B + B)) - B) / B;
 		normalize3(g3[i]);
 	}
 
@@ -207,12 +207,12 @@ void Perlin::init(void)
 }
 
 
-float Perlin::perlin_noise_2D(float vec[2])
+double Perlin::perlin_noise_2D(double vec[2])
 {
 	int terms    = mOctaves;
 	//float freq   = mFrequency;
-	float result = 0.0f;
-	float amp = mAmplitude;
+	double result = 0.0f;
+	double amp = mAmplitude;
 
 	vec[0]*=mFrequency;
 	vec[1]*=mFrequency;
@@ -225,13 +225,12 @@ float Perlin::perlin_noise_2D(float vec[2])
 		amp*=0.5f;
 	}
 
-
 	return result;
 }
 
 
 
-Perlin::Perlin(int octaves,float freq,float amp,int seed)
+Perlin::Perlin(int octaves, double freq, double amp,int seed)
 {
 	mOctaves = octaves;
 	mFrequency = freq;
